pyproject.toml
[tool.poetry]
name = "titanic"
version = "1.0.0"
description = "Machine Learning Engineer Challenge"
authors = ["David E. <daespinosal@unal.edu.co>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.0"
uvicorn = "0.23.2"
pydantic = "^2.0.0"
requests = "^2.32.2"
loguru = "^0.7.0"
joblib = "^1.2.0"
scikit-learn = "^1.5.0"
packaging = "^23.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
httpx = "^0.27.0"
ddtrace = "^2.10.6"
kaggle = "^1.6.17"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
black = {version = "^24.3", allow-prereleases = true}
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
pylint = "^3.0.0"

[tool.poetry.group.aws]
optional = true

[tool.poetry.group.aws.dependencies]
mangum = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''
